refactor(kftray-kube): split start.rs into domain modules
Split kube/start.rs (1065 LOC) into kube/start/:
- mod.rs: entry points (start_port_forward, start_port_forward_with_mode)
- timeout.rs: timeout management (STOPPED_BY_TIMEOUT, callbacks)
- address.rs: loopback address allocation (fallback, IPC, DB persistence)
- forward.rs: single-config forwarding (process_single_config_with_address, TLS, hosts)

Tests travel with their code. Zero behaviour change.
